+++ CHANGELOG.md
@@ -9,8 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improvement for `multisite_operator` ([#722], [#725], [#727]):
   - Add a convenient method when the Hilbert space dimension of all sites are equal.
-  - Improve the documentation and docstring.
-- Efficient Jordan Wigner transformation for `fdestroy` and `fcreate`. ([#723])
+  - Improve the documentation and docstring. 
+- Updates for `fdestroy` and `fcreate`:
+  - Efficient Jordan Wigner transformation. ([#723])
+  - Add Bravyi-Kitaev fermion-to-qubit mapping via the keyword argument `method = :BK` (`method = :JW` for Jordan-Wigner transformation remains the default). ([#724])
 - Fix `mcsolve` error when both keyword arguments `e_ops` and `save_end=false` are specified. ([#728])
 
 ## [v0.47.0]
@@ -525,6 +527,7 @@ Release date: 2024-11-13
 [#718]: https://github.com/qutip/QuantumToolbox.jl/issues/718
 [#722]: https://github.com/qutip/QuantumToolbox.jl/issues/722
 [#723]: https://github.com/qutip/QuantumToolbox.jl/issues/723
+[#724]: https://github.com/qutip/QuantumToolbox.jl/issues/724
 [#725]: https://github.com/qutip/QuantumToolbox.jl/issues/725
 [#727]: https://github.com/qutip/QuantumToolbox.jl/issues/727
 [#728]: https://github.com/qutip/QuantumToolbox.jl/issues/728